moved emulator into its own module and cleaned up emulator.py code
